requirements.txt
biopython>=1.79
pandas>=1.3.0